legacy class seperated from utils, mostly cleaning
