+++ src/contract_clients/mod.rs
@@ -5,3 +5,4 @@ pub mod starknet_sovereign;
 pub mod starknet_validity;
 pub mod token_bridge;
 pub mod utils;
+pub mod legacy_class;
